+++ CodeConverter/CSharp/VBToCSProjectContentsConverter.cs
@@ -43,6 +43,7 @@ namespace ICSharpCode.CodeConverter.CSharp
 
         public async Task InitializeSourceAsync(Project project)
         {
+            project = await ClashingMemberRenamer.RenameClashingSymbolsAsync(project);
             var cSharpCompilationOptions = CSharpCompiler.CreateCompilationOptions();
             _convertedCsProject = project.ToProjectFromAnyOptions(cSharpCompilationOptions, CSharpCompiler.ParseOptions);
             _csharpReferenceProject = project.CreateReferenceOnlyProjectFromAnyOptions(cSharpCompilationOptions, CSharpCompiler.ParseOptions);