+++ Tests/TestData/MultiFileCharacterization/VBToCSResults/ConvertWholeSolution/VbLibrary/AnEnum.cs
@@ -1,7 +1,7 @@
 
 namespace VbLibrary
 {
-    internal enum AnEnum
+    internal enum AnEnumType
     {
         AnEnumMember
     }